requirements.txt
bluepy==1.3.0
paho-mqtt==1.6.1
python-dotenv==0.21.0
environ-config==22.1.0
